Addresses review comments around jobs package, jodatime, autowired dependency moved to int test
