+++ src/test/kotlin/uk/gov/dluhc/printapi/testsupport/testdata/DataBuilder.kt
@@ -1,9 +1,8 @@
 package uk.gov.dluhc.printapi.testsupport.testdata
 
-import org.joda.time.DateTime
-import java.util.UUID
+import java.util.UUID.randomUUID
 
-fun aValidBatchId() = UUID.randomUUID().toString().replace("-", "")
+fun aValidBatchId() = randomUUID().toString().replace("-", "")
 
 fun aValidPrintRequestsFilename() = "05372cf5339447b39f98b248c2217b9f-20221018112232123-10.psv"
 
@@ -13,4 +12,4 @@ fun aValidSftpPath() = "/home/valtech/dev/${aValidZipFilename()}"
 
 fun aValidInputStream() = "Some input stream".byteInputStream()
 
-fun aValidPrintResponseFileName() = "status-" + DateTime.now().toString("yyyyMMddHHmmssSSS") + ".json"
+fun aValidPrintResponseFileName() = "status-20221201171156568.json"
